add module for handling tokens in utils package
